• Gets rid of income values other than sources • renames income reducer
