+++ src/Store/actionTypes.ts
@@ -1,10 +1,10 @@
-import { GlobalIncomeStateType } from "./reducers/income";
+import { GlobalIncomeStateType } from "./reducers/incomeReducer";
 
 type actionType = {
     type: string
 }
 
-export const SET_INCOME_VALUES = "SET_INCOME_VALUES";
-export interface setIncomeValuesActionType extends actionType {
+export const SET_INCOME_SOURCES = "SET_INCOME_SOURCES";
+export interface setIncomeSourcesActionType extends actionType {
     payload: GlobalIncomeStateType
 }
