+++ src/Store/store.ts
@@ -1,6 +1,6 @@
 import { createStore, Store, AnyAction } from "redux";
 import rootReducer from "./reducers";
-import { GlobalIncomeStateType } from "./reducers/income";
+import { GlobalIncomeStateType } from "./reducers/incomeReducer";
 
 export type GlobalStateType = {
     Income: GlobalIncomeStateType;
